random/transport: per-history RNG streams for order-independent parallelism
Implements issue #148 (RNG parallelization Phase 0). Replaces the two
run-wide RNG streams consumed in wavefront order with one independent
stream per particle history, keyed by its global index. A history now
consumes the same random draws regardless of pool capacity, thread, or
rank, removing the coupling between the random sequence and the transport
schedule that blocked reproducible SIMD/threaded/MPI/GPU execution.

RNG API (src/random):
- osh_rng_seed_history(): seed a history's stream from (seed, hist_index,
  purpose) via a SplitMix64 finaliser. Distinct osh_rng_purpose values
  (BEAM, PHYSICS) give independent sub-streams for the same history, so
  source sampling stays decoupled from stochastic transport.
- osh_rng_split(): derive an independent child stream from a parent,
  deterministic along the lineage — the seeding primitive for secondaries.
- struct osh_rng_seeding: carries (type, seed, hist_base) for a run.

Transport / pool:
- The particle pool carries one struct osh_rng per slot; primaries are
  seeded by global history index at fill time, secondaries are split from
  their parent. The stream travels with the particle through compaction.
- prim_idx widened to uint64_t (pool + struct step): it is the seeding key
  and must not alias past 2^32 histories.
- rndoffset is redefined as the global history-index base, so disjoint
  ranges (one per process/MPI rank) yield disjoint, non-overlapping streams.
- Pool capacity is now a runtime knob (osh_simulation_set_pool_capacity /
  --pool-capacity) instead of a compile-time constant; the per-run geometry
  batch buffers are heap-allocated to match (never on the per-step path).

Tests:
- test_osh_rng: order-independence, purpose-independence, disjoint ranges,
  and split determinism for the new seeding primitives.
- bench::capacity_invariance: runs a non-nuclear case across pool
  capacities 1..4096 and asserts scored outputs agree within a tight
  tolerance. The residual (~1e-13, FP summation order in the shared
  scoring accumulators) is the deterministic-reduction follow-up, which
  issue #148 tracks as out of scope.

+++ src/common/osh_step.h
@@ -103,7 +103,7 @@ struct step {
     int medium;        /* material index; -1 if unknown */
     int zone;          /* geometry zone index; -1 if unknown */
     int system;        /* coordinate system (OSH_COORD_*) */
-    uint32_t prim_idx; /* 0-based index of the beam-primary ancestor in the batch */
+    uint64_t prim_idx; /* 0-based index of the beam-primary ancestor in the batch */
     uint8_t gen;       /* generation: 0 = beam primary, 1 = first secondary, … */
     uint8_t has_voxel; /* non-zero when voxel_idx is valid */
 };
